requirements.txt
langgraph
langchain
pydantic
openpyxl
pandas
numpy
tqdm

mcp
fastapi
uvicorn

pillow
pytesseract

tenacity
python-dotenv
openai 